+++ src-tauri/src/lib.rs
@@ -11,9 +11,9 @@ pub fn run() {
         .plugin(tauri_plugin_notification::init())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_shell::init())
+        .plugin(tauri_plugin_process::init())
+        .plugin(tauri_plugin_updater::Builder::new().build())
         .invoke_handler(tauri::generate_handler![
-            get_app_version,
-            check_app_update,
             scan_skills,
             check_github_update,
             update_single_skill,